Speichern von Tags in Gruppe 'tags' (ggf. public schreibbar)
- Erlauben von Mitgliedschaftslöschung schon bei Schreibrecht auf Eltern- und Kindobjekt
- Erstellen/Aktualisieren von Objekten mit mehreren Elternobjekten gleichzeitig

+++ modules/delete.py
@@ -5,13 +5,12 @@ from lib import errors
 errors = imp.reload( errors )
 
 def delete_in( app, object_id_list, parent_id=None ):
-	object_id_list = [ int(x) for x in object_id_list ]
-	for object_id in object_id_list:
-		if not app.user.can_delete( object_id ):
-			raise errors.PrivilegeError( "%d cannot delete %d" % (app.user.id, object_id) )
-	c = app.db.cursor()
 	in_list = ",".join( [str(x) for x in object_id_list] )
+	c = app.db.cursor()
 	if not parent_id:
+		for object_id in object_id_list:
+			if not app.user.can_delete( object_id ):
+				raise errors.PrivilegeError( "%d cannot delete %d" % (app.user.id, object_id) )
 		c.execute( """delete from objects where id in (%(in_list)s)""" % locals() )
 		app.db.commit()
 		# FIXME: Alles folgende ist eigentlich nur nötig, wenn das Backend die 
@@ -23,6 +22,11 @@ def delete_in( app, object_id_list, parent_id=None ):
 		c.execute( """delete from permissions where object_id in (%(in_list)s) or subject_id in (%(in_list)s)""" % locals() )
 		app.db.commit()
 	else:
+		for object_id in object_id_list:
+			if not app.user.can_write( object_id ):
+				raise errors.PrivilegeError( "%d cannot write %d" % (app.user.id, object_id) )
+		if not app.user.can_write( parent_id ):
+			raise errors.PrivilegeError( "%d cannot write %d" % (app.user.id, parent_id) )
 		c.execute( """delete from membership where child_id in (%(in_list)s) and parent_id=?""" % locals(), [parent_id] )
 		app.db.commit()
 
@@ -33,8 +37,8 @@ def process( app ):
 	media_type = None
 	object_id = None
 	if "id" in query.parms:
-		object_id_list = query.parms["id"].split(",")
-		parent_id = "parent_id" in query.parms and query.parms["parent_id"] or None
+		object_id_list = [int(x) for x in query.parms["id"].split(",")]
+		parent_id = "parent_id" in query.parms and int(query.parms["parent_id"]) or None
 		delete_in( app, object_id_list, parent_id=parent_id )
 		response.output = str( {"succeeded" : True, 
 								"delete_id_list" : object_id_list} )
@@ -42,7 +46,7 @@ def process( app ):
 		parent_id = int( query.parms["parent_id"] )
 		object_id_list = []
 		if "child_not_in" in query.parms:
-			object_id_list = query.parms["child_not_in"].split(",")
+			object_id_list = [int(x) for x in query.parms["child_not_in"].split(",")]
 		c = app.db.cursor()
 		not_in_list = ",".join( [str(x) for x in object_id_list] )
 		c.execute( """select child_id from membership where parent_id=? and
